requirements.txt
# Crawl4AI GraphRAG project dependencies

# Crawling / Graph / NLP
crawl4ai[all]>=0.6.0
sentence-transformers>=2.2.0
networkx>=3.0
scikit-learn>=1.3.0
numpy>=1.24.0
summa>=1.2.0
spacy>=3.7.0
pandas>=2.0.0

# Google Gemini integration
google-generativeai>=0.7.0

# Environment & utilities
python-dotenv>=1.0.0
asyncio-run>=0.1.1

# gRPC server & Protocol Buffers
grpcio>=1.73.1
protobuf>=6.31.0 